fix(test): OCP Tests are failing on MacOS
Code review updates,
- using contextlib from the standard library so no updates to
  requirements are necessary.
- Moving the multiprocess context manager out of the
  specific OpenShift scan tests and to the DummyScanManager.
  This allows the fix to work with other users of the
  DummyScanManager.
