+++ requirements.in
@@ -1,5 +1,4 @@
 ansible>=6,<7
-contextlib2>=21.6.0
 Django<4
 django-filter
 djangorestframework
